pyproject.toml
[project]
name = "flink-mcp"
version = "0.1.0"
description = "MCP server for Apache Flink SQL Gateway"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
authors = [{ name = "Maciek Brynski", email = "maciej.brynski@cledar.com" }]

[project.scripts]
flink-mcp = "flink_mcp.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["flink_mcp*"]

[tool.setuptools.package-data]
flink_mcp = ["py.typed"]

[project.license]
text = "Apache-2.0"

[project.urls]
Homepage = "https://github.com/cledar/flink-mcp"
Repository = "https://github.com/cledar/flink-mcp"
Issues = "https://github.com/cledar/flink-mcp/issues"


[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: tests that hit a live Flink SQL Gateway (set SQL_GATEWAY_API_BASE_URL)",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
